+++ src/fractal_task_tools/_generatejsonschema.py
@@ -1,12 +1,15 @@
 import logging
 from typing import Any
 
+from pydantic import BaseModel
 from pydantic.json_schema import GenerateJsonSchema
 from pydantic.json_schema import JsonSchemaValue
 from pydantic.json_schema import NoDefault
 from pydantic_core import core_schema
 
 logger = logging.getLogger("CustomGenerateJsonSchema")
+
+BaseModel.model_config["use_attribute_docstrings"] = True
 
 
 class CustomGenerateJsonSchema(GenerateJsonSchema):