+++ tests/fake-tasks/src/fake_tasks/task_list.py
@@ -3,10 +3,8 @@ from fractal_task_tools.task_models import NonParallelTask
 AUTHORS = "Fake Fake"
 DOCS_LINK = "https://www.example.org"
 INPUT_MODELS = [
-    ["fake_tasks", "task1.py", "MyModel"],
+    # ("fake_tasks", "models.py", "ModelMixedDocstrings")
 ]
-
-
 TASK_LIST = [
     NonParallelTask(
         name="Task1",